+++ libs/api/a-guarantees/src/lib/controllers/guarantees.controller.ts
@@ -1,5 +1,6 @@
+import { Roles, RolesGuard } from '@ivt/a-auth';
 import { GuaranteeEntity } from '@ivt/a-state';
-import { GuaranteeSummary } from '@ivt/c-data';
+import { GuaranteeSummary, UserRoles } from '@ivt/c-data';
 import {
   Body,
   Controller,
@@ -23,7 +24,7 @@ import { UpdateGuaranteeDto } from '../dto/update-dtos/update-guarantee.dto';
 import { GuaranteesService } from '../services/guarantees.service';
 
 @Controller('guarantees')
-@UseGuards(AuthGuard())
+@UseGuards(AuthGuard(), RolesGuard)
 export class GuaranteesController {
   constructor(private guaranteesService: GuaranteesService) {}
 
@@ -42,18 +43,27 @@ export class GuaranteesController {
     return this.guaranteesService.getGuaranteesSummary();
   }
 
+  @Get('export/excel')
+  async getGuaranteesExcel(
+    @Query(ValidationPipe) filterDto: GetGuaranteesFilterDto,
+    @Res() response: Response
+  ): Promise<void> {
+    return this.guaranteesService.getGuaranteesExcel(filterDto, response);
+  }
+
   @Post()
   @UsePipes(new ValidationPipe({ transform: true }))
   async createGuarantee(@Body() createGuaranteeDto: CreateGuaranteeDto) {
     return this.guaranteesService.createGuarantee(createGuaranteeDto);
   }
 
-  @Get(':id/pdf')
-  async getGuaranteePdf(@Param('id', ParseIntPipe) id: number, @Res() response: Response) {
+  @Get('export/pdf/:id')
+  async getGuaranteePdf(@Param('id', ParseIntPipe) id: number, @Res() response: Response): Promise<void> {
     return this.guaranteesService.generatePdf(id, response);
   }
 
   @Put(':id')
+  @Roles(UserRoles.superAdmin)
   @UsePipes(new ValidationPipe({ transform: true }))
   updateGuarantee(@Body() updateGuaranteeDto: UpdateGuaranteeDto): Promise<GuaranteeEntity> {
     return this.guaranteesService.updateGuarantee(updateGuaranteeDto);