+++ libs/common/c-data/src/lib/core/enums/http-status-codes.enum.ts
@@ -1,0 +1,6 @@
+export enum HttpStatusCodes {
+  Unauthorized = 401,
+  BadRequest = 402,
+  NotFound = 404,
+  InternalServerError = 500,
+}