+++ libs/common/c-data/src/lib/core/models/http-error-response.model.ts
@@ -1,5 +1,7 @@
+import { HttpStatusCodes } from '../enums';
+
 export interface IvtHttpErrorResponse {
   error: string;
   message: string;
-  statusCode: string;
+  statusCode: HttpStatusCodes;
 }