+++ libs/common/c-data/src/lib/users/enums/user-roles.enum.ts
@@ -1,0 +1,5 @@
+export enum UserRoles {
+  superAdmin = 1,
+  admin = 2,
+  agencyUser = 3,
+}