+++ libs/ui/u-state/src/lib/entities/index.ts
@@ -1,3 +1,4 @@
+export * from './build-query-params';
 export * from './entity-collection-reducer-methods';
 export * from './entity-collection-reducer-methods.factory';
 export * from './entity.metadata';