+++ libs/ui/u-state/src/lib/permissions/directives/create-permission.directive.spec.ts
@@ -1,0 +1,8 @@
+import { CreatePermissionDirective } from './create-permission.directive';
+
+describe('CreatePermissionDirective', () => {
+  it('should create an instance', () => {
+    const directive = new CreatePermissionDirective();
+    expect(directive).toBeTruthy();
+  });
+});