+++ libs/ui/u-state/src/lib/permissions/services/permission.service.ts
@@ -1,0 +1,29 @@
+import { Injectable, InjectionToken } from '@angular/core';
+import { UserRoles } from '@ivt/c-data';
+import { filterNil } from '@ivt/c-utils';
+import { select, Store } from '@ngrx/store';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
+
+import { getAuthUserRoleState } from '../../auth';
+import { PermissionTypes } from '../enums/permission-types.enum';
+
+export const REQUIRED_PERMISSIONS = new InjectionToken<PermissionTypes[]>(
+  'Required Permission Types which the user must have to complete an action (CRUD privileges)'
+);
+
+@Injectable({
+  providedIn: 'root',
+})
+export class PermissionService {
+  userRole$ = this.store.pipe(select(getAuthUserRoleState));
+
+  constructor(private store: Store<any>) {}
+
+  hasUpdatePermission(...requiredPermmissionTypes: PermissionTypes[]): Observable<boolean> {
+    return this.userRole$.pipe(
+      filterNil(),
+      map(role => role === UserRoles[UserRoles.superAdmin])
+    );
+  }
+}