+++ libs/ui/u-ui/src/lib/components/form/form.component.ts
@@ -1,12 +1,7 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  EventEmitter,
-  Input,
-  Output,
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 
+import { markFormGroupTouched } from '../../functions';
 import { IvtSubscriberComponent } from '../subscriber';
 
 @Component({
@@ -18,6 +13,7 @@ export class IvtFormComponent<T = any> extends IvtSubscriberComponent {
   @Input() form: FormGroup;
   @Input() item: Partial<T>;
   @Input() loading: boolean;
+  @Input() isEditable = true;
   @Output() submitForm = new EventEmitter<T>();
 
   get values(): T {
@@ -28,7 +24,7 @@ export class IvtFormComponent<T = any> extends IvtSubscriberComponent {
     if (this.form.valid || this.form.disabled) {
       this.submitForm.emit(this.values);
     } else {
-      this.form.markAllAsTouched();
+      markFormGroupTouched(this.form);
     }
   }
 }