+++ libs/ui/u-ui/src/lib/components/table/table-header/table-header.component.ts
@@ -1,12 +1,6 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  EventEmitter,
-  Input,
-  Output,
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 
-import { IvtColumns, SortEvent } from '../table.model';
+import { IvtBreakpointShow, IvtColumns, IvtColumnSizes, SortEvent } from '../table.model';
 
 @Component({
   selector: 'ivt-table-header',
@@ -22,5 +16,22 @@ export class IvtTableHeaderComponent {
   updateSort(sortEvent: SortEvent): void {
     this.activeFilter = sortEvent.value;
     this.sort.emit(sortEvent);
+  }
+
+  getClasses(columnSizes: IvtColumnSizes, show: IvtBreakpointShow): any {
+    return {
+      col: !columnSizes?.xs && !columnSizes?.sm && !columnSizes?.md && !columnSizes?.lg && !columnSizes?.xl,
+      [`col-${columnSizes?.xs}`]: !!columnSizes?.xs,
+      [`col-sm-${columnSizes?.sm}`]: !!columnSizes?.sm,
+      [`col-md-${columnSizes?.md}`]: !!columnSizes?.md,
+      [`col-lg-${columnSizes?.lg}`]: !!columnSizes?.lg,
+      [`col-xl-${columnSizes?.xl}`]: !!columnSizes?.xl,
+      [`d-none`]: !!show,
+      [`d-xs-block`]: !!show?.xs,
+      [`d-sm-block`]: !!show?.sm,
+      [`d-md-block`]: !!show?.md,
+      [`d-lg-block`]: !!show?.lg,
+      [`d-xl-block`]: !!show?.xl,
+    };
   }
 }