+++ libs/ui/u-ui/src/lib/containers/list-container/list-container.component.ts
@@ -1,9 +1,9 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { IvtQueryParams } from '@ivt/c-data';
-import { IvtEntityCollection, IvtCollectionService } from '@ivt/u-state';
 import { filterNil } from '@ivt/c-utils';
+import { buildQueryParams, IvtCollectionService, IvtDataService, IvtEntityCollection } from '@ivt/u-state';
 import { select } from '@ngrx/store';
-import { map, takeUntil } from 'rxjs/operators';
+import { map, take, takeUntil } from 'rxjs/operators';
 
 import { IvtSubscriberComponent } from '../../components';
 
@@ -20,22 +20,17 @@ export class IvtListContainerComponent<T = any> extends IvtSubscriberComponent {
     filterNil(),
     map((collection: IvtEntityCollection) => collection.hasMore)
   );
+  loadingExcel$ = this.entityDataService.loadingExcel$;
   queryParams: IvtQueryParams;
+  excelFileName: string;
+  excelUrl: string;
 
-  constructor(
-    protected entityCollectionService: IvtCollectionService<T>
-  ) {
+  constructor(protected entityCollectionService: IvtCollectionService<T>, protected entityDataService: IvtDataService) {
     super();
     this.entityCollectionService.store
-      .pipe(
-        select(this.entityCollectionService.selectors.selectCollection),
-        filterNil(),
-        takeUntil(this.destroy$)
-      )
-      .subscribe(
-        (collection: IvtEntityCollection) =>
-          (this.queryParams = collection.queryParams)
-      );
+      .pipe(select(this.entityCollectionService.selectors.selectCollection), filterNil(), takeUntil(this.destroy$))
+      .subscribe((collection: IvtEntityCollection) => (this.queryParams = collection.queryParams));
+    this.excelUrl = `${this.entityDataService.getEntitiesUrl()}/export/excel`;
   }
 
   getMoreItems(): void {
@@ -46,15 +41,23 @@ export class IvtListContainerComponent<T = any> extends IvtSubscriberComponent {
     this.entityCollectionService.getWithQuery(queryParams as any);
   }
 
-  filterItems<TQueryParams = any>(
-    payload: IvtQueryParams & Partial<TQueryParams>
-  ): void {
+  filterItems<TQueryParams = any>(payload: IvtQueryParams & Partial<TQueryParams>): void {
     const queryParams: IvtQueryParams = {
       ...this.queryParams,
       ...payload,
       resetList: true,
     };
     this.entityCollectionService.getWithQuery(queryParams as any);
+  }
+
+  exportExcel(queryParams?: IvtQueryParams): void {
+    const payload = {
+      params: buildQueryParams({ ...queryParams, ...this.queryParams }),
+      url: this.excelUrl,
+      fileName: this.excelFileName,
+    };
+
+    this.entityDataService.exportExcel(payload).pipe(take(1)).subscribe();
   }
 
   deleteItem(item: T): void {