+++ libs/ui/u-ui/src/lib/pipes/folio/folio.module.ts
@@ -7,5 +7,6 @@ import { IvtFolioPipe } from './folio.pipe';
   declarations: [IvtFolioPipe],
   imports: [CommonModule],
   exports: [IvtFolioPipe],
+  providers: [IvtFolioPipe]
 })
 export class IvtFolioModule {}